fix: Fix bug in iOS; screen appears blank when modal opens (#538)
* Create static scrollTo method that checks if window is defined

* Create scrollPosition state

* Scroll window to top, set scrollPosition state when modal state is open

This is done to fix a bug seen in iOS Chrome and Safari where the
on-screen keyboard ruins the scroll position of the modal when
opened and appears as if the screen is blank.

When the modal is open, noScroll is turned on, which calculates the
current scroll position and sets this value to html.style.top. This
value is transformed and set to the scrollPosition state. Lastly,
the scrollTo method is used to scroll the window to top and prevent
the keyboard from causing the scroll bar to miscalculate the window
height.

* Scroll to scrollPosition state when the modal open state is false

When the modal is closed, the window is scrolled back to where the
user was before the modal was opened. This creates a better experience
and prevents an annoying scroll jump when the modal is closed. This is
just the value that is calculated by noScroll.on() and stored in the
state when the modal opens.