feat: Create ProfileIntro component
This component has been pulled out of ProfileHeader.

It features a read more / read less action to expand / collapse the text
which uses react-measure to calculate the height and line clamping to
truncate text.

And because legacy user profile bios allowed for Markdown,
react-markdown-renderer is used to render this content.