feat: Switch Bookmark and BookmarkAlt icons, deprecate BookmarkAlt
Since the bookmark icons with the plus and checkmark are no longer being
used, they have been renamed "Alt" and have been deprecated. The icons
that were previously identified as alternate have now been renamed and
are now the primary bookmark icons.

This also updates the paths for Bookmark and BookmarkActive icons.