feat: Refactor SectionHeader component
* Replace media queries with react-measure. This allows for the
component’s size to dictate its responsive styles instead of the
window’s size. In turn, the component has been converted from
functional to stateful because the dimensions provided by react-measure
need to be handled within the component’s state.
* Remove containing `header` element. There is no reason for an extra
container around the Heading component, so it has been removed.
* Use new TextHeading component and typography styles
* Change divider from `hr` to a `span`. Ideally, the divider would be
build purely in CSS using pseudo elements; this change mimics that
idea. 1. A pseudo element lives inside of it’s parent element, so the
divider markup has been moved inside of Heading; this also allows for
the sizing to be relative to the Heading’s font size. 2. A pseudo
element has no semantic value, neither does a `span` and since the
divider is presentational, the `hr` provides semantic value we don’t
want.
* Remove heading object prop; since the component itself is now a
Heading, we don’t need a separate prop to pass in an object of settings
for the Heading. Furthermore, the Heading now has it’s props
predetermined, i.e., size, level, weight. Previously, the
SectionHeading could include any number of heading props, but now I
think it’s better to limit the component to be specifically configured
how it was designed; an h2, using our fontSizeHeading2 (4 for mobile)
and regular weight. This will allow the component to be rendered
consistently.
* These updates should not break any current implementation, however,
any heading settings passed in will not be applied.