feat: Update newsletter component (#546)
* Send JSON data, use XHR
* Update UI and improve accessibility

lonelyplanet/wp-travelnews-frontend#344

* feat: Delete Recaptcha site key, pass it in with a prop

* fix: Remove label prop from form/input

The input's label shouldn't be placed on the element itself with a title
or aria-label attribute. The label should use the label element.

* fix: Remove extra whitespace in form/input

* fix: Use propTypes.style for form/input

* fix: Refactor how styles are applied in form/input

* feat: Create keyCode mapping utility

* feat: Improve accessibility of checkbox component

There was a significant bug where the checkbox element could not be
tabbed to or selected with the keyboard.

A native HTML checkbox can be toggled when focused with the spacebar
key. This update adds that functionality back in.

This component itself may need to be re-thought and re-built from
an accessibility perspective, for not only keyboard users, but also
for screen readers.

* chore: Ignore .env file from git

* fix: Add endpoint, captchaSiteKey props to Newsletter story which use env vars

* fix: Throw error in Newsletter component if captcha site key is missing

* fix: Make captchaSiteKey prop required in Newsletter component

* feat: Add endpoint prop to Newsletter component

* fix: Update UI, styles for Newsletter component

* feat: Send data from Newsletter component as JSON

* fix: Update newsletter component with changes from #565